requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
google-api-python-client
pytrends
